Separation of concerns by cearting a component for individual items and pass it down as child of TaskList component
